refactor(routing): Enhance route protection and error handling
