+++ frontend/src/router/components/protected-route.tsx
@@ -1,16 +1,15 @@
-import { lazy, useCallback, useEffect } from 'react';
+import React, { useCallback, useEffect } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
 
+import PageError from '@/pages/sys/error/PageError';
 import { useUserToken } from '@/store/user';
 
 import { useRouter } from '../hooks';
 
-const PageError = lazy(() => import('@/pages/sys/error/PageError'));
-
 type Props = {
   children: React.ReactNode;
 };
-export default function AuthGuard({ children }: Props) {
+export default function ProtectedRoute({ children }: Props) {
   const router = useRouter();
   const { accessToken } = useUserToken();
