+++ frontend/src/router/routes/error-routes.tsx
@@ -3,10 +3,9 @@ import { Outlet } from 'react-router-dom';
 
 import { CircleLoading } from '@/components/Loading';
 import SimpleLayout from '@/layouts/SimpleLayout';
+import ProtectedRoute from '@/router/components/protected-route';
 
-import AuthGuard from '../components/auth-guard';
-
-import { AppRouteObject } from '#/router';
+import type { AppRouteObject } from '#/router';
 
 const Page403 = lazy(() => import('@/pages/sys/error/Page403'));
 const Page404 = lazy(() => import('@/pages/sys/error/Page404'));
@@ -16,15 +15,15 @@ const Page500 = lazy(() => import('@/pages/sys/error/Page500'));
  * error routes
  * 403, 404, 500
  */
-export const ErrorRoutes: AppRouteObject = {
+export const ERROR_ROUTE: AppRouteObject = {
   element: (
-    <AuthGuard>
+    <ProtectedRoute>
       <SimpleLayout>
         <Suspense fallback={<CircleLoading />}>
           <Outlet />
         </Suspense>
       </SimpleLayout>
-    </AuthGuard>
+    </ProtectedRoute>
   ),
   children: [
     { path: '403', element: <Page403 /> },
